+++ src/payload/model.rs
@@ -140,7 +140,7 @@ pub struct Issue {
     created_at: OffsetDateTime,
     #[serde(with = "time::serde::rfc3339")]
     updated_at: OffsetDateTime,
-    #[serde(with = "option_time_rfc3339")]
+    #[serde(with = "super::option_time_rfc3339")]
     closed_at: Option<OffsetDateTime>,
     body: Option<String>,
 }
@@ -210,31 +210,4 @@ pub struct Changes {
 #[get = "pub"]
 pub struct ChangedFrom {
     from: String,
-}
-
-// This trick is originally provided by @dtolnay at:
-// https://github.com/serde-rs/serde/issues/1301#issuecomment-394108486
-/// A helper module to workaround serde with customize functions for `Option`
-/// value.
-mod option_time_rfc3339 {
-    use serde::{Deserialize, Deserializer, Serialize, Serializer};
-    use time::OffsetDateTime;
-
-    pub fn serialize<S: Serializer>(
-        value: &Option<OffsetDateTime>,
-        serializer: S,
-    ) -> Result<S::Ok, S::Error> {
-        #[derive(Serialize)]
-        struct Wrapper<'a>(#[serde(with = "time::serde::rfc3339")] &'a OffsetDateTime);
-        value.as_ref().map(Wrapper).serialize(serializer)
-    }
-
-    pub fn deserialize<'de, D: Deserializer<'de>>(
-        deserializer: D,
-    ) -> Result<Option<OffsetDateTime>, D::Error> {
-        #[derive(Deserialize)]
-        struct Wrapper(#[serde(with = "time::serde::rfc3339")] OffsetDateTime);
-        let wrapper = Option::deserialize(deserializer)?;
-        Ok(wrapper.map(|Wrapper(datetime)| datetime))
-    }
 }
