+++ tests/testdata/star_deleted.json
@@ -1,0 +1,124 @@
+{
+  "action": "deleted",
+  "starred_at": null,
+  "repository": {
+    "id": 455788009,
+    "node_id": "R_kgDOGyrF6Q",
+    "name": "testdev",
+    "full_name": "tisonkun/testdev",
+    "private": false,
+    "owner": {
+      "login": "tisonkun",
+      "id": 18818196,
+      "node_id": "MDQ6VXNlcjE4ODE4MTk2",
+      "avatar_url": "https://avatars.githubusercontent.com/u/18818196?v=4",
+      "gravatar_id": "",
+      "url": "https://api.github.com/users/tisonkun",
+      "html_url": "https://github.com/tisonkun",
+      "followers_url": "https://api.github.com/users/tisonkun/followers",
+      "following_url": "https://api.github.com/users/tisonkun/following{/other_user}",
+      "gists_url": "https://api.github.com/users/tisonkun/gists{/gist_id}",
+      "starred_url": "https://api.github.com/users/tisonkun/starred{/owner}{/repo}",
+      "subscriptions_url": "https://api.github.com/users/tisonkun/subscriptions",
+      "organizations_url": "https://api.github.com/users/tisonkun/orgs",
+      "repos_url": "https://api.github.com/users/tisonkun/repos",
+      "events_url": "https://api.github.com/users/tisonkun/events{/privacy}",
+      "received_events_url": "https://api.github.com/users/tisonkun/received_events",
+      "type": "User",
+      "site_admin": false
+    },
+    "html_url": "https://github.com/tisonkun/testdev",
+    "description": null,
+    "fork": false,
+    "url": "https://api.github.com/repos/tisonkun/testdev",
+    "forks_url": "https://api.github.com/repos/tisonkun/testdev/forks",
+    "keys_url": "https://api.github.com/repos/tisonkun/testdev/keys{/key_id}",
+    "collaborators_url": "https://api.github.com/repos/tisonkun/testdev/collaborators{/collaborator}",
+    "teams_url": "https://api.github.com/repos/tisonkun/testdev/teams",
+    "hooks_url": "https://api.github.com/repos/tisonkun/testdev/hooks",
+    "issue_events_url": "https://api.github.com/repos/tisonkun/testdev/issues/events{/number}",
+    "events_url": "https://api.github.com/repos/tisonkun/testdev/events",
+    "assignees_url": "https://api.github.com/repos/tisonkun/testdev/assignees{/user}",
+    "branches_url": "https://api.github.com/repos/tisonkun/testdev/branches{/branch}",
+    "tags_url": "https://api.github.com/repos/tisonkun/testdev/tags",
+    "blobs_url": "https://api.github.com/repos/tisonkun/testdev/git/blobs{/sha}",
+    "git_tags_url": "https://api.github.com/repos/tisonkun/testdev/git/tags{/sha}",
+    "git_refs_url": "https://api.github.com/repos/tisonkun/testdev/git/refs{/sha}",
+    "trees_url": "https://api.github.com/repos/tisonkun/testdev/git/trees{/sha}",
+    "statuses_url": "https://api.github.com/repos/tisonkun/testdev/statuses/{sha}",
+    "languages_url": "https://api.github.com/repos/tisonkun/testdev/languages",
+    "stargazers_url": "https://api.github.com/repos/tisonkun/testdev/stargazers",
+    "contributors_url": "https://api.github.com/repos/tisonkun/testdev/contributors",
+    "subscribers_url": "https://api.github.com/repos/tisonkun/testdev/subscribers",
+    "subscription_url": "https://api.github.com/repos/tisonkun/testdev/subscription",
+    "commits_url": "https://api.github.com/repos/tisonkun/testdev/commits{/sha}",
+    "git_commits_url": "https://api.github.com/repos/tisonkun/testdev/git/commits{/sha}",
+    "comments_url": "https://api.github.com/repos/tisonkun/testdev/comments{/number}",
+    "issue_comment_url": "https://api.github.com/repos/tisonkun/testdev/issues/comments{/number}",
+    "contents_url": "https://api.github.com/repos/tisonkun/testdev/contents/{+path}",
+    "compare_url": "https://api.github.com/repos/tisonkun/testdev/compare/{base}...{head}",
+    "merges_url": "https://api.github.com/repos/tisonkun/testdev/merges",
+    "archive_url": "https://api.github.com/repos/tisonkun/testdev/{archive_format}{/ref}",
+    "downloads_url": "https://api.github.com/repos/tisonkun/testdev/downloads",
+    "issues_url": "https://api.github.com/repos/tisonkun/testdev/issues{/number}",
+    "pulls_url": "https://api.github.com/repos/tisonkun/testdev/pulls{/number}",
+    "milestones_url": "https://api.github.com/repos/tisonkun/testdev/milestones{/number}",
+    "notifications_url": "https://api.github.com/repos/tisonkun/testdev/notifications{?since,all,participating}",
+    "labels_url": "https://api.github.com/repos/tisonkun/testdev/labels{/name}",
+    "releases_url": "https://api.github.com/repos/tisonkun/testdev/releases{/id}",
+    "deployments_url": "https://api.github.com/repos/tisonkun/testdev/deployments",
+    "created_at": "2022-02-05T05:53:04Z",
+    "updated_at": "2022-02-06T12:48:00Z",
+    "pushed_at": "2022-02-05T05:53:04Z",
+    "git_url": "git://github.com/tisonkun/testdev.git",
+    "ssh_url": "git@github.com:tisonkun/testdev.git",
+    "clone_url": "https://github.com/tisonkun/testdev.git",
+    "svn_url": "https://github.com/tisonkun/testdev",
+    "homepage": null,
+    "size": 0,
+    "stargazers_count": 0,
+    "watchers_count": 0,
+    "language": null,
+    "has_issues": true,
+    "has_projects": true,
+    "has_downloads": true,
+    "has_wiki": true,
+    "has_pages": false,
+    "forks_count": 0,
+    "mirror_url": null,
+    "archived": false,
+    "disabled": false,
+    "open_issues_count": 1,
+    "license": null,
+    "allow_forking": true,
+    "is_template": false,
+    "topics": [
+
+    ],
+    "visibility": "public",
+    "forks": 0,
+    "open_issues": 1,
+    "watchers": 0,
+    "default_branch": "main"
+  },
+  "sender": {
+    "login": "tisonkun",
+    "id": 18818196,
+    "node_id": "MDQ6VXNlcjE4ODE4MTk2",
+    "avatar_url": "https://avatars.githubusercontent.com/u/18818196?v=4",
+    "gravatar_id": "",
+    "url": "https://api.github.com/users/tisonkun",
+    "html_url": "https://github.com/tisonkun",
+    "followers_url": "https://api.github.com/users/tisonkun/followers",
+    "following_url": "https://api.github.com/users/tisonkun/following{/other_user}",
+    "gists_url": "https://api.github.com/users/tisonkun/gists{/gist_id}",
+    "starred_url": "https://api.github.com/users/tisonkun/starred{/owner}{/repo}",
+    "subscriptions_url": "https://api.github.com/users/tisonkun/subscriptions",
+    "organizations_url": "https://api.github.com/users/tisonkun/orgs",
+    "repos_url": "https://api.github.com/users/tisonkun/repos",
+    "events_url": "https://api.github.com/users/tisonkun/events{/privacy}",
+    "received_events_url": "https://api.github.com/users/tisonkun/received_events",
+    "type": "User",
+    "site_admin": false
+  }
+}
